separate Condition Clauses into files
